pyproject.toml
[project]
name = "co-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo==0.19.11",
    "mcp>=1.0.0",
    "pandas>=3.0.0",
    "pytest>=8.0.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
